# [type] Add BitStructStoreStmt and corresponding passes