--- a/misc/benchmark_bit_struct_stores.py
+++ b/misc/benchmark_bit_struct_stores.py
@@ -7,7 +7,7 @@ quant = True
 n = 1024 * 1024 * 256
 
 if quant:
-    ci16 = ti.type_factory_.get_custom_int_type(16, True)
+    ci16 = ti.type_factory.custom_int(16, True)
 
     x = ti.field(dtype=ci16)
     y = ti.field(dtype=ci16)