--- a/python/taichi/lang/type_factory.py
+++ b/python/taichi/lang/type_factory.py
@@ -3,8 +3,8 @@ class TypeFactory:
         from taichi.core import ti_core
         self.core = ti_core.get_type_factory_instance()
 
-    def custom_int(self, bits, signed=True):
-        return self.core.get_custom_int_type(bits, signed)
+    def custom_int(self, bits, signed=True, compute_bits=0):
+        return self.core.get_custom_int_type(bits, signed, compute_bits)
 
     def custom_float(self,
                      significand_type,