--- a/taichi/codegen/codegen_llvm.cpp
+++ b/taichi/codegen/codegen_llvm.cpp
@@ -347,8 +347,7 @@ void CodeGenLLVM::visit(UnaryOpStmt *stmt) {
     } else if (!is_real(from) && !is_real(to)) {
       // TODO: implement casting into custom integer type
       TI_ASSERT(!to->is<CustomIntType>());
-      llvm_val[stmt] =
-          cast_int(llvm_val[stmt->operand], from.get_ptr(), to.get_ptr());
+      llvm_val[stmt] = cast_int(llvm_val[stmt->operand], from, to);
     }
   } else if (stmt->op_type == UnaryOpType::cast_bits) {
     TI_ASSERT(data_type_size(stmt->ret_type) ==
@@ -1029,8 +1028,7 @@ llvm::Value *CodeGenLLVM::atomic_add_custom_int(AtomicOpStmt *stmt,
       fmt::format("atomic_add_partial_bits_b{}", data_type_bits(physical_type)),
       {builder->CreateBitCast(byte_ptr, llvm_ptr_type(physical_type)),
        bit_offset, tlctx->get_constant(cit->get_num_bits()),
-       cast_int(llvm_val[stmt->val], stmt->val->ret_type.get_ptr(),
-                physical_type)});
+       cast_int(llvm_val[stmt->val], stmt->val->ret_type, physical_type)});
 }
 
 llvm::Value *CodeGenLLVM::atomic_add_custom_float(AtomicOpStmt *stmt,
@@ -1291,7 +1289,7 @@ void CodeGenLLVM::visit(BitStructStoreStmt *stmt) {
     for (int i = 0; i < stmt->ch_ids.size(); i++) {
       auto ch_id = stmt->ch_ids[i];
       auto val = llvm_val[stmt->values[i]];
-      auto dtype = bit_struct_snode->ch[ch_id]->dt.get_ptr();
+      auto dtype = bit_struct_snode->ch[ch_id]->dt;
       val = custom_type_to_bits(val, dtype, bit_struct_physical_type);
       val = builder->CreateShl(val, bit_struct_snode->ch[ch_id]->bit_offset);
       if (bit_struct_val == nullptr) {
@@ -1315,10 +1313,9 @@ void CodeGenLLVM::visit(BitStructStoreStmt *stmt) {
       } else {
         cit = dtype->as<CustomIntType>();
       }
-      store_custom_int(llvm_val[stmt->ptr], tlctx->get_constant(ch->bit_offset),
-                       cit,
-                       custom_type_to_bits(llvm_val[val], dtype.get_ptr(),
-                                           bit_struct_physical_type));
+      store_custom_int(
+          llvm_val[stmt->ptr], tlctx->get_constant(ch->bit_offset), cit,
+          custom_type_to_bits(llvm_val[val], dtype, bit_struct_physical_type));
     }
   }
 }
@@ -1615,7 +1612,7 @@ void CodeGenLLVM::visit(SNodeLookupStmt *stmt) {
     llvm_val[stmt] = parent;
   } else if (snode->type == SNodeType::bit_array) {
     auto element_num_bits =
-        snode->dt.get_ptr()->as<BitArrayType>()->get_element_num_bits();
+        snode->dt->as<BitArrayType>()->get_element_num_bits();
     auto offset = tlctx->get_constant(element_num_bits);
     offset = builder->CreateMul(offset, llvm_val[stmt->input_index]);
     llvm_val[stmt] = create_bit_ptr_struct(llvm_val[stmt->input_snode], offset);
@@ -1629,7 +1626,7 @@ void CodeGenLLVM::visit(GetChStmt *stmt) {
   if (stmt->input_snode->type == SNodeType::bit_array) {
     llvm_val[stmt] = llvm_val[stmt->input_ptr];
   } else if (stmt->ret_type->as<PointerType>()->is_bit_pointer()) {
-    auto bit_struct = stmt->input_snode->dt.get_ptr()->cast<BitStructType>();
+    auto bit_struct = stmt->input_snode->dt->cast<BitStructType>();
     auto bit_offset = bit_struct->get_member_bit_offset(
         stmt->input_snode->child_id(stmt->output_snode));
     auto offset = tlctx->get_constant(bit_offset);