--- a/taichi/ir/type_factory.cpp
+++ b/taichi/ir/type_factory.cpp
@@ -40,6 +40,9 @@ Type *TypeFactory::get_pointer_type(Type *element, bool is_bit_pointer) {
 Type *TypeFactory::get_custom_int_type(int num_bits,
                                        bool is_signed,
                                        int compute_type_bits) {
+  if (compute_type_bits == 0) {
+    compute_type_bits = 32;
+  }
   auto key = std::make_tuple(compute_type_bits, num_bits, is_signed);
   if (custom_int_types.find(key) == custom_int_types.end()) {
     custom_int_types[key] = std::make_unique<CustomIntType>(