--- a/taichi/llvm/llvm_context.cpp
+++ b/taichi/llvm/llvm_context.cpp
@@ -588,6 +588,8 @@ llvm::Value *TaichiLLVMContext::get_constant(DataType dt, T t) {
 
 template llvm::Value *TaichiLLVMContext::get_constant(DataType dt, int32 t);
 template llvm::Value *TaichiLLVMContext::get_constant(DataType dt, int64 t);
+template llvm::Value *TaichiLLVMContext::get_constant(DataType dt, uint32 t);
+template llvm::Value *TaichiLLVMContext::get_constant(DataType dt, uint64 t);
 template llvm::Value *TaichiLLVMContext::get_constant(DataType dt, float32 t);
 template llvm::Value *TaichiLLVMContext::get_constant(DataType dt, float64 t);