--- a/taichi/llvm/llvm_context.cpp
+++ b/taichi/llvm/llvm_context.cpp
@@ -573,14 +573,14 @@ llvm::Value *TaichiLLVMContext::get_constant(DataType dt, T t) {
     return llvm::ConstantFP::get(*ctx, llvm::APFloat((float32)t));
   } else if (dt->is_primitive(PrimitiveTypeID::f64)) {
     return llvm::ConstantFP::get(*ctx, llvm::APFloat((float64)t));
-  } else if (dt->is_primitive(PrimitiveTypeID::i32)) {
-    return llvm::ConstantInt::get(*ctx, llvm::APInt(32, t, true));
-  } else if (dt->is_primitive(PrimitiveTypeID::u32)) {
-    return llvm::ConstantInt::get(*ctx, llvm::APInt(32, t, false));
-  } else if (dt->is_primitive(PrimitiveTypeID::i64)) {
-    return llvm::ConstantInt::get(*ctx, llvm::APInt(64, t, true));
-  } else if (dt->is_primitive(PrimitiveTypeID::u64)) {
-    return llvm::ConstantInt::get(*ctx, llvm::APInt(64, t, false));
+  } else if (is_integral(dt)) {
+    if (is_signed(dt)) {
+      return llvm::ConstantInt::get(*ctx,
+                                    llvm::APInt(data_type_bits(dt), t, true));
+    } else {
+      return llvm::ConstantInt::get(*ctx,
+                                    llvm::APInt(data_type_bits(dt), t, false));
+    }
   } else {
     TI_NOT_IMPLEMENTED
   }