--- a/taichi/python/export_lang.cpp
+++ b/taichi/python/export_lang.cpp
@@ -89,11 +89,13 @@ void export_lang(py::module &m) {
       .def(py::self == py::self)
       .def("__hash__", &DataType::hash)
       .def("to_string", &DataType::to_string)
-      .def("get_ptr", &DataType::get_ptr, py::return_value_policy::reference)
+      .def("get_ptr", [](const DataType &dtype) -> Type * { return dtype; },
+           py::return_value_policy::reference)
       .def(py::pickle(
           [](const DataType &dt) {
             // Note: this only works for primitive types, which is fine for now.
-            auto primitive = dynamic_cast<const PrimitiveType *>(dt.get_ptr());
+            auto primitive =
+                dynamic_cast<const PrimitiveType *>((const Type *)dt);
             TI_ASSERT(primitive);
             return py::make_tuple((std::size_t)primitive->type);
           },