--- a/taichi/python/export_lang.cpp
+++ b/taichi/python/export_lang.cpp
@@ -89,7 +89,7 @@ void export_lang(py::module &m) {
       .def(py::self == py::self)
       .def("__hash__", &DataType::hash)
       .def("to_string", &DataType::to_string)
-      .def("get_ptr", [](const DataType &dtype) -> Type * { return dtype; },
+      .def("get_ptr", [](DataType *dtype) -> Type * { return *dtype; },
            py::return_value_policy::reference)
       .def(py::pickle(
           [](const DataType &dt) {