@xumingkuan 2020-12-29
I'm not sure how will `BitStructStoreStmt` affect `data_sourse_analysis.cpp` and other potential IR optimizations assuming `GlobalStoreStmt` and `AtomicOpStmt` are the only statements modifying global fields. However, we are good now as long as this pass is after all other optimizations.